backend/requirements.txt
# Backend requirements for RAG Chatbot
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=0.19.0
transformers>=4.21.0
torch>=1.13.0
scikit-learn>=1.1.0
numpy>=1.21.0
requests>=2.30.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
openai>=1.0.0